% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{find_Lineage}
\alias{find_Lineage}
\title{By an explicit taxon name, all taxonomic lineages under that taxno are found.}
\usage{
find_Lineage(nodename, level = 1)
}
\arguments{
\item{nodename}{the name of the taxno.}
\item{level}{a number, that is the number of levels you want to expand, i.e. how many levels of categorical relationships you want to obtain for all the lineages under this definite node. A maximum of 5 levels can be expanded. The default is 1.}
}
\value{
Finally, return a list.`[[1]]`is a S3 class for the phylogenetic tree. `[[2]]`is a dataframe showing the correspondence between parent and child nodes
}
\description{
By an explicit taxon name, all taxonomic lineages under that taxno are found.
}
\references{
\enumerate{
\item Hadley Wickham.
\url{https://github.com/tidyverse/tidyverse}
\item G Yu, DK Smith, H Zhu, Y Guan, TTY Lam (2017). ggtree: an R package for
visualization and annotation of phylogenetic trees with their covariates and
other associated data. Methods in Ecology and Evolution, 8(1):28-36.
\url{https://doi.org/10.1111/2041-210X.12628}
}

Please, when using taxtree, cite us using the reference:
\url{https://github.com/nongxinshengxin/taxtree}
}
\author{
Cy G
}