% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{name2rank}
\alias{name2rank}
\title{Get the taxonomy rank name based on the name of the taxa.}
\usage{
name2rank(file, header = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.This file contains the names of all the taxa for which you want to build a classification topology. Each line is a separate taxon name.}
\item{header}{a logical value indicating whether the file contains the names of the variables as its first line. If missing, the value is determined from the file format: header is set to TRUE if and only if the first row contains one fewer field than the number of columns. Default is FALSE}
}
\description{
If you have some definite taxa names (either Kingdom Phylum Class Order Family Genus Species or any other taxonomic node), you can use this function to get the taxonomy rank name based on the name of the taxa.
}
\references{
\url{https://github.com/nongxinshengxin/taxtree}
}
