% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{name2rank_str}
\alias{name2rank_str}
\title{Get the taxonomy rank name based on the name of the taxa.(String vector as input)}
\usage{
name2rank_str(nodenames)
}
\arguments{
\item{nodenames}{the names of all the taxa for which you want to build a classification topology. You can input a single string or a vector containing multiple strings.}
}
\description{
If you have some definite taxa names (either Kingdom Phylum Class Order Family Genus Species or any other taxonomic node), you can use this function to get the taxonomy rank name based on the name of the taxa.
}
\references{
\url{https://github.com/nongxinshengxin/taxtree}
}
