% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names.dmp}
\alias{names.dmp}
\title{names.dmp}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2354163 rows and 2 columns.
}
\usage{
names.dmp
}
\description{
Taxid and species name correspondence file
}
\examples{
  head(names.dmp)
}
\keyword{datasets}
