% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{nodes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2354163 rows and 3 columns.
}
\usage{
nodes
}
\description{
Species Hierarchy Document
}
\examples{
  head(nodes)
}
\keyword{datasets}
