% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{plot_taxTree}
\alias{plot_taxTree}
\title{Drawing a simple Taxonomy tree based on the `ggtree` package}
\usage{
plot_taxTree(
  treefile,
  module = 1，
  clade_group = c())
}
\arguments{
\item{treefile}{phylo object}
\item{clade_group}{grouping clades, input a vector containing strings}
\item{module}{the layout of tree plot, with three types, 1, 2 and 3, corresponding to 'rectangular', 'ellipse' and 'circular'}
}
\description{
Drawing a simple Taxonomy tree based on the `ggtree` package.
}
\references{
\url{https://github.com/nongxinshengxin/taxtree}
}
