% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{write_taxTree}
\alias{write_taxTree}
\title{Write Tree File in Parenthetic Format}
\usage{
write_taxTree(tree, output)
}
\arguments{
\item{tree}{an object of class "phylo".}
\item{output}{a file name specified by either a variable of mode character, or a double-quoted string.}
}
\description{
This function writes in a file a tree in parenthetic format using the Newick format, based on `ape` packages
}
\references{
\enumerate{
\item Hadley Wickham.
\url{https://github.com/tidyverse/tidyverse}
\item G Yu, DK Smith, H Zhu, Y Guan, TTY Lam (2017). ggtree: an R package for
visualization and annotation of phylogenetic trees with their covariates and
other associated data. Methods in Ecology and Evolution, 8(1):28-36.
\url{https://doi.org/10.1111/2041-210X.12628}
}

Please, when using taxtree, cite us using the reference:
\url{https://github.com/nongxinshengxin/taxtree}
}
\author{
Cy G
}
