% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{find_Lineage}
\alias{find_Lineage}
\title{By an explicit taxonomic node name, all taxonomic lineages under that node are found.}
\usage{
find_Lineage(nodename, level = 3)
}
\arguments{
\item{nodename}{the name of the node.}
}
\item{level}{a number, that is the number of levels you want to expand, i.e. how many levels of categorical relationships you want to obtain for all the lineages under this definite node. A maximum of 5 levels can be expanded. The default is 3.}
}
\description{
By an explicit taxonomic node name, all taxonomic lineages under that node are found.
}
\references{
\url{https://github.com/nongxinshengxin/taxtree}
}
