% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{make_Taxtree}
\alias{make_Taxtree}
\title{Construct their categorised topology from the list of taxa names.}
\usage{
make_Taxtree(file, header = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.This file contains the names of all the taxa for which you want to build a taxonomic topology. Each line is a separate taxon name.}
\item{header}{a logical value indicating whether the file contains the names of the variables as its first line. If missing, the value is determined from the file format: header is set to TRUE if and only if the first row contains one fewer field than the number of columns. Default is FALSE}
}
\description{
If you have some definite taxa names (either Kingdom Phylum Class Order Family Genus Species or any other taxonomic node), you can use this function to construct their taxonomic topology from a list of taxa names.
}
\references{
\enumerate{
\item Hadley Wickham.
\url{https://github.com/tidyverse/tidyverse}
\item G Yu, DK Smith, H Zhu, Y Guan, TTY Lam (2017). ggtree: an R package for
visualization and annotation of phylogenetic trees with their covariates and
other associated data. Methods in Ecology and Evolution, 8(1):28-36.
\url{https://doi.org/10.1111/2041-210X.12628}
}

Please, when using taxtree, cite us using the reference:
\url{https://github.com/nongxinshengxin/taxtree}
}
\author{
Cy G
}
