% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{plot_taxTree}
\alias{plot_taxTree}
\title{Drawing a simple Taxonomy tree based on the `ggtree` package}
\usage{
plot_taxTree(
  treefile,
  module = 1，
  clade_group = c())
}
\arguments{
\item{treefile}{phylo object}
\item{clade_group}{grouping clades, input a vector containing strings}
\item{module}{the layout of tree plot, with three types, 1, 2 and 3, corresponding to 'rectangular', 'ellipse' and 'circular'}
}
\description{
Drawing a simple Taxonomy tree based on the `ggtree` package.
}
\references{
\enumerate{
\item Hadley Wickham.
\url{https://github.com/tidyverse/tidyverse}
\item G Yu, DK Smith, H Zhu, Y Guan, TTY Lam (2017). ggtree: an R package for
visualization and annotation of phylogenetic trees with their covariates and
other associated data. Methods in Ecology and Evolution, 8(1):28-36.
\url{https://doi.org/10.1111/2041-210X.12628}
}

Please, when using taxtree, cite us using the reference:
\url{https://github.com/nongxinshengxin/taxtree}
}
\author{
Cy G
}
