% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{write_taxTree}
\alias{write_taxTree}
\title{Write Tree File in Parenthetic Format}
\usage{
write_taxTree(tree, output)
}
\arguments{
\item{tree}{an object of class "phylo".}
\item{output}{a file name specified by either a variable of mode character, or a double-quoted string.}
}
\description{
This function writes in a file a tree in parenthetic format using the Newick format, based on `ape` packages
}
\references{
\url{https://github.com/nongxinshengxin/taxtree}
}
